+++ scrapy/http/request/json_request.py
@@ -1,5 +1,5 @@
 """
-This module implements the JSONRequest class which is a more convenient class
+This module implements the JsonRequest class which is a more convenient class
 (than Request) to generate JSON Requests.
 
 See documentation in docs/topics/request-response.rst
@@ -12,7 +12,7 @@ import warnings
 from scrapy.http.request import Request
 
 
-class JSONRequest(Request):
+class JsonRequest(Request):
     def __init__(self, *args, **kwargs):
         dumps_kwargs = copy.deepcopy(kwargs.pop('dumps_kwargs', {}))
         dumps_kwargs.setdefault('sort_keys', True)
@@ -31,7 +31,7 @@ class JSONRequest(Request):
             if 'method' not in kwargs:
                 kwargs['method'] = 'POST'
 
-        super(JSONRequest, self).__init__(*args, **kwargs)
+        super(JsonRequest, self).__init__(*args, **kwargs)
         self.headers.setdefault('Content-Type', 'application/json')
         self.headers.setdefault('Accept', 'application/json, text/javascript, */*; q=0.01')
 
@@ -46,7 +46,7 @@ class JSONRequest(Request):
         elif not body_passed and data_passed:
             kwargs['body'] = self._dumps(data)
 
-        return super(JSONRequest, self).replace(*args, **kwargs)
+        return super(JsonRequest, self).replace(*args, **kwargs)
 
     def _dumps(self, data):
         """Convert to JSON """